// src/interfaces/Card.ts
export type Card = ManualGeneratedFields & AutoGeneratedFields;

interface ManualGeneratedFields {
  title: string;
  question: string;
  answer: string;
  tags: string[];
  images: string[];
}

interface AutoGeneratedFields {
  _id: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateCardModel = ManualGeneratedFields;